Share generated text sanitization and strip commit spam
- Move attribution stripping into `packages/shared`
- Add commit-msg hook to remove provider attribution from commit messages
- Expand tests and cleanup duplicate PR utility exports
