+++ .husky/commit-msg
@@ -1,0 +1,1 @@
+bun run scripts/strip-provider-attribution-from-commit-message.ts "$1"
