+++ apps/web/src/components/pr-review/pr-review-utils.tsx
@@ -21,10 +21,6 @@ import { normalizeLanguageIdForHighlighting } from "~/lib/languageIds";
 
 export { parseRenderablePatch, resolveFileDiffPath, summarizeFileDiffStats };
 
-export { parseRenderablePatch, resolveFileDiffPath, summarizeFileDiffStats };
-
-export { parseRenderablePatch, resolveFileDiffPath, summarizeFileDiffStats };
-
 export type PullRequestState = "open" | "closed" | "merged";
 export type InspectorTab = "threads" | "workflow" | "people";
 export type RequestChangesButtonVariant = "default" | "destructive-outline" | "outline";
